refactor: move functions to lib.rs
# Context
Previously all files were in main.rs. They have been
moved to lib.rs for modularity.
