+++ src/main.rs
@@ -1,33 +1,7 @@
 use std::env;
-use std::error::Error;
-use std::fs;
 use std::process::exit;
 
-struct Config {
-    query: String,
-    file_path: String,
-}
-
-impl Config {
-    fn build(args: &[String]) -> Result<Config, &'static str> {
-        if args.len() < 3 {
-            return Err("Not enough arguments!");
-        }
-
-        let query = args[1].clone();
-        let file_path = args[2].clone();
-
-        Ok(Config { query, file_path })
-    }
-}
-
-fn run(config: Config) -> Result<(), Box<dyn Error>> {
-    let contents = fs::read_to_string(config.file_path)?;
-
-    print!("With text:\n{contents}");
-
-    Ok(())
-}
+use minigrep::Config;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -40,7 +14,7 @@ fn main() {
     println!("Searching for: {}", config.query);
     println!("In file: {}", config.file_path);
 
-    if let Err(e) = run(config) {
+    if let Err(e) = minigrep::run(config) {
         println!("Application error: {e}");
         exit(1)
     }
